+++ compiler/src/iree/compiler/Preprocessing/Common/ApplyPDLPatterns.cpp
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "PDLUtilities.h"
+#include "Utilities.h"
 #include "iree/compiler/Preprocessing/Common/Passes.h"
 
 #include "iree/compiler/Codegen/Dialect/Codegen/IR/IREECodegenDialect.h"
