+++ compiler/src/iree/compiler/Preprocessing/Common/BUILD.bazel
@@ -36,13 +36,13 @@ iree_compiler_cc_library(
         "ConvertConvToChannelsLast.cpp",
         "InterpreterPass.cpp",
         "MakeSingleDispatchForFunction.cpp",
-        "PDLUtilities.cpp",
+        "Utilities.cpp",
         "PadLinalgOps.cpp",
         "PadToIntrinsics.cpp",
         "Passes.cpp",
     ],
     hdrs = [
-        "PDLUtilities.h",
+        "Utilities.h",
         "Passes.h",
         "Passes.h.inc",
     ],
